feat: Adding responsive behaviour for the images of the site
Add attributes to all images files for optimizing their use.
The image resize to three differents sizes, 260, 460 y 800 px wide.
Create a grunt script for automate the creating images in different
sizes.

A few adjust in javascript files for create the img srcset and sizes
attributes and change the table structure for better adjust of the
timetable table of the restaurant details view for get the split hours
in differents lines.